requirements.txt
django
psycopg2
flake8
pep8
coverage
python-coveralls
pillow
factory-boy
django-widget-tweaks